FDG-6154 Relocate accordion
