+++ src/layouts/explainer/sections/government-revenue/federal-revenue-trends-and-us-economy/federal-revenue-trends-and-us-economy.jsx
@@ -3,12 +3,12 @@ import TotalRevenueChart
   from "./government-revenue-and-us-economy-chart/total-revenue-chart/total-revenue-chart";
 import useBeaGDP from "../../../../../hooks/useBeaGDP";
 import { faMapLocationDot } from '@fortawesome/free-solid-svg-icons';
-import Accordion from "../../../../../components/accordion/accordion";
 import CustomLink from "../../../../../components/links/custom-link/custom-link";
 import QuoteBox from "../../../quote-box/quote-box";
-import {revenueExplainerPrimary, revenueExplainerLightSecondary, revenueAccordion}
+import {revenueExplainerPrimary, revenueExplainerLightSecondary}
   from "../revenue.module.scss";
 import {quoteBoxContent} from "../../../explainer.module.scss";
+import {section} from "./federal-revenue-trends-and-us-economy.module.scss";
 
 
 const FederalRevenueTrendsAndUSEconomy = ({cpiDataByYear}) => {
@@ -24,11 +24,6 @@ const FederalRevenueTrendsAndUSEconomy = ({cpiDataByYear}) => {
     setRevenueTotal(data.revenueTotal);
   }
 
-  const federalReserveAct = (
-    <CustomLink url={'https://www.federalreserve.gov/aboutthefed/section7.htm'} >
-      Federal Reserve Act, Section 7(a)(1-3)
-    </CustomLink>
-  );
 
   const gps = (
     <CustomLink url={'https://www.gps.gov/policy/funding/'} >
@@ -37,7 +32,7 @@ const FederalRevenueTrendsAndUSEconomy = ({cpiDataByYear}) => {
   );
 
   return (
-    <>
+    <div className={section}>
       <p>
         In fiscal year {fiscalYear}, federal revenue was
         equal to {revenueRatio} of total gross domestic product (GDP), or economic
@@ -69,22 +64,7 @@ const FederalRevenueTrendsAndUSEconomy = ({cpiDataByYear}) => {
           <span style={{ fontSize: "16px" }}>Source: {gps}</span>
         </p>
       </QuoteBox>
-      <div className={revenueAccordion}>
-        <Accordion
-          title={'Why does the Federal Reserve send money to the federal government?'}
-          openEventNumber={"13"}
-          explainerGAEvent={"Revenue"}
-        >
-          The Federal Reserve Act of 1913, the law that created the Federal Reserve System and
-          established it as the central bank of the United States, requires that any excess
-          earnings generated by 12 Federal Reserve Banks be paid to the federal government.
-          In practice, Federal Reserve banks pay for their operating expenses, pay a dividend
-          to stockholders (U.S. commercial banks), set aside money in a surplus fund (capped at
-          $6.8 billion), and the remainder is “transferred…to the Secretary of the Treasury for
-          deposit in the general fund of the Treasury.” [<i>See</i> {federalReserveAct}]
-        </Accordion>
-      </div>
-    </>
+    </div>
   );
 };
