+++ src/layouts/explainer/sections/government-revenue/federal-revenue-trends-and-us-economy/federal-revenue-trends-and-us-economy.module.scss
@@ -1,10 +1,6 @@
 @import '../../../../../variables.module.scss';
 
-
-.comingSoon {
-  padding: 2px;
-  font-style: italic;
-  font-size: $font-size-16;
-
+.section {
+  margin-bottom: 2rem;
 }
