+++ src/layouts/explainer/sections/government-revenue/sources-of-federal-revenue/sources-of-federal-revenue.jsx
@@ -11,10 +11,12 @@ import {
 } from "../revenue.module.scss";
 import { sourcesContent } from "./sources-of-federal-revenue.module.scss";
 import { faMartiniGlassCitrus } from "@fortawesome/free-solid-svg-icons";
-import SourcesOfRevenueCircleChart from "./sources-of-revenue-circle-chart/sources-of-revenue-circle-chart";
+import SourcesOfRevenueCircleChart
+  from "./sources-of-revenue-circle-chart/sources-of-revenue-circle-chart";
 import { apiPrefix, basicFetch } from "../../../../../utils/api-utils";
 import GlossaryTerm from "../../../../../components/glossary-term/glossary-term";
-
+import Accordion from "../../../../../components/accordion/accordion";
+import {revenueAccordion} from "../revenue.module.scss";
 const SourcesOfFederalRevenue = ({ glossary }) => {
   const [currentFiscalYear, setCurrentFiscalYear] = useState(0);
   const [indvPercent, setIndvPercent] = useState('');
@@ -103,6 +105,12 @@ const SourcesOfFederalRevenue = ({ glossary }) => {
     </GlossaryTerm>
   );
 
+  const federalReserveAct = (
+    <CustomLink url={'https://www.federalreserve.gov/aboutthefed/section7.htm'} >
+      Federal Reserve Act, Section 7(a)(1-3)
+    </CustomLink>
+  );
+
   return (
     <div className={sourcesContent}>
       <p>
@@ -123,6 +131,21 @@ const SourcesOfFederalRevenue = ({ glossary }) => {
       </p>
       <div className={visWithCallout}>
         <SourcesOfRevenueCircleChart />
+      </div>
+      <div className={revenueAccordion}>
+        <Accordion
+          title={'Why does the Federal Reserve send money to the federal government?'}
+          openEventNumber={"13"}
+          explainerGAEvent={"Revenue"}
+        >
+          The Federal Reserve Act of 1913, the law that created the Federal Reserve System and
+          established it as the central bank of the United States, requires that any excess
+          earnings generated by 12 Federal Reserve Banks be paid to the federal government.
+          In practice, Federal Reserve banks pay for their operating expenses, pay a dividend
+          to stockholders (U.S. commercial banks), set aside money in a surplus fund (capped at
+          $6.8 billion), and the remainder is “transferred…to the Secretary of the Treasury for
+          deposit in the general fund of the Treasury.” [<i>See</i> {federalReserveAct}]
+        </Accordion>
       </div>
       <h5>Social Security and Medicare Taxes</h5>
       <p>
